note update route

+++ pages/api/note/[id].ts
@@ -5,7 +5,22 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
+  const { title, content } = req.body;
   const noteId = req.query.id;
+
+  if (req.method === "POST") {
+    const updatedNote = await prisma.note.update({
+      where: {
+        id: Number(noteId),
+      },
+      data: {
+        title,
+        content,
+      },
+    });
+
+    return res.status(200).json({ message: "succesfully updated note" });
+  }
 
   if (req.method === "DELETE") {
     const note = await prisma.note.delete({
@@ -13,6 +28,6 @@ export default async function handler(
     });
     return res.json(noteId);
   } else {
-    return res.json({ message: "couldn't delete note" });
+    return res.json({ message: "error" });
   }
 }
